pyproject.toml
[tool.poetry]
name = "tickets-reservation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tickets_reservation"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
drf-nested-routers = "^0.93.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
flake8 = "^6.0.0"
black = "^22.12.0"
flake8-pyproject = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
)/
'''

[tool.flake8]
max-line-length = 120
extend-ignore = ["E402"]
exclude = [".git", "__pycache__", "dist"]
max-complexity = 10

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true